requirements.txt
Flask
python-dotenv
google-generativeai
 Flask
+gunicorn
 python-dotenv
 google-generativeai
